requirements.txt
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
